implement delete account call

+++ findaroo-frontend/src/Pages/Profile.js
@@ -4,6 +4,8 @@ import GlobalVariables from "../Utils/GlobalVariables";
 import {Link, useNavigate} from "react-router-dom";
 import {useEffect, useState} from "react";
 import ButtonStandard, {ButtonDelete, ButtonImportant} from "../Components/Buttons";
+import { confirmAlert } from 'react-confirm-alert';
+import 'react-confirm-alert/src/react-confirm-alert.css';
 
 export default function Profile() {
     // This redirects to the login page if not logged in
@@ -114,16 +116,21 @@ export default function Profile() {
     }
 
     async function DeleteAccountCall(userId) {
-        userId = "abs";
+        let answer = window.confirm("Your record will be removed from Findaroo. Are you sure?");
+        if (!answer) return;
         const response = await fetch(GlobalVariables.backendURL + "/User", {
             mode: 'cors',
             method: 'DELETE',
             headers: {
-                'Content-type': 'application/json; charset=UTF-8'
+                'Content-type': 'application/json; charset=UTF-8',
+                'Access-Control-Allow-Origin': 'http://localhost:3000'
             },
-            body: JSON.stringify(userId)
+            body: JSON.stringify({"user_id": userId})
         }).then(response => setUserData(response.json()))
             .catch(error => console.error(error));
+
+        GlobalVariables.authenticated = false;
+        navigate("/Login");
     }
 
     /*
